@@ -9,11 +9,10 @@
   "dependencies": {
     "dayjs": "^1.10.3",
     "framer-motion": "^7.0.0",
-    "next": "^12.1.0",
+    "next": "^13.0.0",
     "node-html-parser": "^6.0.0",
     "react": "^18.0.0",
     "react-dom": "^18.0.0",
-    "react-ga4": "^1.4.1",
-    "swr": "^1.0.0"
+    "react-ga4": "^1.4.1"
   }
 }
