@@ -1,17 +1,23 @@
+'use client'
+
 import Link from 'next/link'
-import { useRouter } from 'next/router'
+import { useSelectedLayoutSegments } from 'next/navigation'
 import themes from './themes'
 
 import styles from '../styles/Nav.module.css'
 
 const Nav = () => {
-    const { asPath } = useRouter()
+    const [selectedLayoutSegment] = useSelectedLayoutSegments()
 
     return (
         <header className={styles.header}>
             {themes.map((theme, i) => (
-                <Link href='/[theme]' as={theme.link} key={i}>
-                    <a className={asPath === theme.link ? styles.current : ''}>{theme.title}</a>
+                <Link 
+                    href={`/${theme.link}`} 
+                    key={i} 
+                    className={selectedLayoutSegment === theme.link ? styles.current : ''}
+                >
+                    {theme.title}
                 </Link>
             ))}
         </header>
