@@ -1,10 +1,10 @@
 const themes = [
-    {title: "World",    link: "/world",         color: "#35b789"},
-    {title: "Россия",   link: "/nation",        color: "#ff4f40"},
-    {title: "Спорт",    link: "/sports",        color: "#88CCF1"},
-    {title: "Наука",    link: "/scitech",       color: "#FFB86F"},
-    {title: "Культура", link: "/entertainment", color: "#DCED31"},
-    {title: "Бизнес",   link: "/business",      color: "#91A6FF"}
+    {title: "World",    link: "world",         color: "#35b789"},
+    {title: "Россия",   link: "nation",        color: "#ff4f40"},
+    {title: "Спорт",    link: "sports",        color: "#88CCF1"},
+    {title: "Наука",    link: "scitech",       color: "#FFB86F"},
+    {title: "Культура", link: "entertainment", color: "#DCED31"},
+    {title: "Бизнес",   link: "business",      color: "#91A6FF"}
 ]
 
 export default themes
